requirements.txt
Flask>=2.0.0
pyodbc
gunicorn
python-dotenv>=0.21.0